+++ pkg/webui/console/views/application-collaborators-list/index.js
@@ -31,6 +31,7 @@ import {
   selectCollaborators,
   selectCollaboratorsTotalCount,
   selectCollaboratorsFetching,
+  selectCollaboratorsError,
 } from '@console/store/selectors/collaborators'
 
 export default class ApplicationCollaborators extends React.Component {
@@ -54,6 +55,7 @@ export default class ApplicationCollaborators extends React.Component {
       collaborators: selectCollaborators(state, id),
       fetching: selectCollaboratorsFetching(state),
       totalCount: selectCollaboratorsTotalCount(state, id),
+      error: selectCollaboratorsError(state),
     }
   }