+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Setting custom gateway attributes in the console
+- Setting custom gateway attributes in the Console.
 - Pub/Sub documentation.
 - Return informative well-known errors for standard network and context errors.
 - Skipping payload crypto on application-level via application link's `skip_payload_crypto` field.
+- Error notification in list views in the Console.
+- Latest "last seen" info and uplink frame counts for end devices in the Console.
+- Latest "last seen" info for applications in the Console.
 
 ### Changed